Doing "ReactApp - Tic Tac Toe" - Fix the code and files filling Wolox React style guide.
