+++ react/react-training/src/screens/Game.css
