+++ react/react-training/src/screens/MyScreenComponent/Square.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+function Square(props) {
+  return (
+    <button className="square" onClick={props.onClick}>
+      {props.value}
+    </button>
+  );
+}
+
+Square.propTypes = {
+  onClick: PropTypes.string.isRequired,
+  value: PropTypes.string.isRequired
+};
+
+export default Square;
